+++ avalanche/evaluation/metrics/cpu_usage.py
@@ -22,7 +22,7 @@ from avalanche.evaluation.metric_utils import get_metric_name, \
 from avalanche.evaluation.metrics import Mean
 
 if TYPE_CHECKING:
-    from avalanche.training.plugins import PluggableStrategy
+    from avalanche.training import PluggableStrategy
 
 
 class CPUUsage(Metric[float]):
