+++ avalanche/evaluation/metrics/disk_usage.py
@@ -19,7 +19,7 @@ from avalanche.evaluation.metric_utils import get_metric_name, \
 from avalanche.evaluation.metric_results import MetricResult, MetricValue
 
 if TYPE_CHECKING:
-    from avalanche.training.plugins import PluggableStrategy
+    from avalanche.training import PluggableStrategy
 
 PathAlike = Union[Union[str, Path]]
