+++ avalanche/evaluation/metrics/loss.py
@@ -20,7 +20,7 @@ from avalanche.evaluation.metric_utils import get_metric_name, \
     phase_and_task, stream_type
 from avalanche.evaluation.metrics.mean import Mean
 if TYPE_CHECKING:
-    from avalanche.training.plugins import PluggableStrategy
+    from avalanche.training import PluggableStrategy
 
 
 class Loss(Metric[float]):
