+++ avalanche/evaluation/metrics/mac.py
@@ -19,7 +19,7 @@ from avalanche.evaluation.metric_utils import get_metric_name, \
     phase_and_task, stream_type
 
 if TYPE_CHECKING:
-    from avalanche.training.plugins import PluggableStrategy
+    from avalanche.training import PluggableStrategy
 
 
 class MAC(Metric[int]):
