+++ avalanche/training/__init__.py
@@ -5,3 +5,6 @@ CL strategies. These are provided either as standalone strategies in
 :py:mod:`training.strategies` or as plugins (:py:mod:`training.plugins`) that
 can be easily combined with your own strategy.
 """
+from typing import Union
+
+PluggableStrategy = Union['BaseStrategy', 'JointTraining']
