+++ avalanche/training/plugins/__init__.py
@@ -1,0 +1,12 @@
+from .agem import AGEMPlugin
+from .cwr_star import CWRStarPlugin
+from .evaluation import EvaluationPlugin
+from .ewc import EWCPlugin
+from .gdumb import GDumbPlugin
+from .gem import GEMPlugin
+from .lwf import LwFPlugin
+from .multi_head import MultiHeadPlugin
+from .replay import ReplayPlugin
+from .strategy_plugin import StrategyPlugin
+from .synaptic_intelligence import SynapticIntelligencePlugin
+
